requirements.txt
APScheduler==3.6.3
arrow==0.15.5
asgiref==3.2.3
blessed==1.17.1
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.3
django-picklefield==2.1.1
django-q==1.1.0
future==0.18.2
gunicorn==20.0.4
idna==2.8
line-bot-sdk==1.15.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
static3==0.7.0
tzlocal==2.0.0
urllib3==1.25.8
wcwidth==0.1.8
psycopg2==2.7.3.2
